Implements most of the expression engine except user functions.

+++ src/propan/frontend/ast.zig
@@ -69,7 +69,7 @@ pub const Expression = union(enum) {
 pub const IntegerLiteral = struct {
     location: Location,
     source_text: []const u8,
-    value: u64,
+    value: u32,
 };
 
 pub const StringLiteral = struct {
