+++ src/propan/frontend/parser.zig
@@ -520,21 +520,21 @@ pub const Parser = struct {
             return output.toOwnedSlice(allocator);
         }
 
-        fn parse_int(core: *Core, text: []const u8) !u64 {
+        fn parse_int(core: *Core, text: []const u8) !u32 {
             _ = core;
             if (std.mem.startsWith(u8, text, "0b"))
-                return try std.fmt.parseInt(u64, text[2..], 2);
+                return try std.fmt.parseInt(u32, text[2..], 2);
 
             if (std.mem.startsWith(u8, text, "0q"))
-                return try std.fmt.parseInt(u64, text[2..], 4);
+                return try std.fmt.parseInt(u32, text[2..], 4);
 
             if (std.mem.startsWith(u8, text, "0o"))
-                return try std.fmt.parseInt(u64, text[2..], 8);
+                return try std.fmt.parseInt(u32, text[2..], 8);
 
             if (std.mem.startsWith(u8, text, "0x"))
-                return try std.fmt.parseInt(u64, text[2..], 16);
+                return try std.fmt.parseInt(u32, text[2..], 16);
 
-            return try std.fmt.parseInt(u64, text, 10);
+            return try std.fmt.parseInt(u32, text, 10);
         }
 
         // if(!C & !Z)`, `if(>)`
